+++ src/ActionListeners/Admin/ConfermaInserimentoCategoriaListener.java
@@ -1,4 +1,4 @@
-package ActionListeners;
+package ActionListeners.Admin;
 
 import Business.InserimentoCategoriaBusiness;
 import DAO.CategoriaDAO;
