+++ src/ActionListeners/Admin/ConfermaInserimentoFornitoreListener.java
@@ -1,4 +1,4 @@
-package ActionListeners;
+package ActionListeners.Admin;
 
 import Business.InserimentoFornitoreBusiness;
 import DAO.FornitoreDAO;
