+++ src/ActionListeners/Admin/ConfermaInserimentoManagerListener.java
@@ -1,4 +1,4 @@
-package ActionListeners;
+package ActionListeners.Admin;
 
 import Business.InserimentoManagerBusiness;
 import DAO.Interfaces.IUtenteDAO;
