+++ src/ActionListeners/Admin/ConfermaInserimentoProduttoriListener.java
@@ -1,4 +1,4 @@
-package ActionListeners;
+package ActionListeners.Admin;
 
 import Business.InserimentoProduttoreBusiness;
 import DAO.Interfaces.IProduttoreDAO;
