+++ src/ActionListeners/Admin/CreazioneFotoListener.java
@@ -1,4 +1,4 @@
-package ActionListeners;
+package ActionListeners.Admin;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
