+++ src/ActionListeners/Admin/CreazioneProdottoListener.java
@@ -1,15 +1,15 @@
-package ActionListeners;
+package ActionListeners.Admin;
 
 import Views.FinestraPrincipale;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class GestioneUtentiListener implements ActionListener {
-    public final static String GESTISCIUTENTI_BTN = "GestisciUtenti_btn";
+public class CreazioneProdottoListener implements ActionListener {
+    public final static String INSERISCIPRODOTTO_BTN = "InserisciProdotto_btn";
     private FinestraPrincipale frame;
 
-    public GestioneUtentiListener(FinestraPrincipale frame) {
+    public CreazioneProdottoListener(FinestraPrincipale frame) {
         this.frame = frame;
     }
 
@@ -20,9 +20,8 @@ public class GestioneUtentiListener implements ActionListener {
     @Override
     public void actionPerformed(ActionEvent e) {
         String azione = e.getActionCommand();
-
-        if (GESTISCIUTENTI_BTN.equals(azione)) {
-            frame.mostraPannelloGestioneUtenti();
+        if (INSERISCIPRODOTTO_BTN.equals(azione)) {
+            frame.mostraPannelloInserimentoProdotto();
         }
     }
 }
