+++ src/ActionListeners/Manager/GestioneUtentiListener.java
@@ -1,15 +1,15 @@
-package ActionListeners;
+package ActionListeners.Manager;
 
 import Views.FinestraPrincipale;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class CreazioneProdottoListener implements ActionListener {
-    public final static String INSERISCIPRODOTTO_BTN = "InserisciProdotto_btn";
+public class GestioneUtentiListener implements ActionListener {
+    public final static String GESTISCIUTENTI_BTN = "GestisciUtenti_btn";
     private FinestraPrincipale frame;
 
-    public CreazioneProdottoListener(FinestraPrincipale frame) {
+    public GestioneUtentiListener(FinestraPrincipale frame) {
         this.frame = frame;
     }
 
@@ -20,8 +20,9 @@ public class CreazioneProdottoListener implements ActionListener {
     @Override
     public void actionPerformed(ActionEvent e) {
         String azione = e.getActionCommand();
-        if (INSERISCIPRODOTTO_BTN.equals(azione)) {
-            frame.mostraPannelloInserimentoProdotto();
+
+        if (GESTISCIUTENTI_BTN.equals(azione)) {
+            frame.mostraPannelloGestioneUtenti();
         }
     }
 }
