+++ src/ActionListeners/Manager/RispondiCommentiListener.java
@@ -1,4 +1,4 @@
-package ActionListeners;
+package ActionListeners.Manager;
 
 import Views.FinestraPrincipale;
